+++ o365spray/core/handlers/enumerator/modules/office.py
@@ -98,8 +98,7 @@ class EnumerateModule_office(EnumeratorBase):
                 if_exists_result = int(body["IfExistsResult"])
                 is_request_throttled = int(body["ThrottleStatus"])
 
-                # Check if the requests are being throttled and shutdown
-                # if so
+                # Check if the requests are being throttled for the given user
                 if is_request_throttled != 0 or if_exists_result == 2:
                     logging.info(f"Requests are being throttled for: {email}")