+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update `office` enumeration module to continue past request throttling as it is per user
 - Remove spraying modules that use Basic Authentication (ActiveSync, Autodiscover, Reporting) due to [Microsoft's Deprecation of Basic Authentication](https://learn.microsoft.com/en-us/exchange/clients-and-mobile-in-exchange-online/deprecation-of-basic-authentication-exchange-online)
 - Remove the sleep and jitter during domain validation
+- Add support for FireProx handling directly through the tool via fire.py
 
 ## v3.0.4 (11/19/2023)
 - [@Macmod](https://github.com/Macmod) Update how enumeration performs concurrent requests by limiting the number of blocking tasks at any one time